Update DEPLOYMENT-GUIDE.md
This is a fantastic solution accelerator! Josh Bradley and the other contributors really helped me understand how to build these solutions. 

The GRAPHRAG_API_VERSION variable in the deploy.parameters.json file can be more clearly defined in the deployment readme. I made the mistake of accidentally using the model version here as it looked similar to the expected value and my work colleagues made the same exact mistake as they were also bringing the solution online. The failure I was experiencing was that the indexing was not failing, but running up to 12.5% and then going back to 6.25% and cycling. As soon as I changed API Version to 2024-05-01-preview and redeployed, everything worked!

It was the errors in the K8 logs (openai.NotFoundError: Error code: 404 Resource not found) that pointed me to review the Azure OpenAI API Version and I got everything working last night. 

I absolutely love this project and will be trying to actively contribute to the development. 

I also found a lot of value monitoring the Azure OpenAI Metrics Dashboard from the Azure Portal so we can see how the HTTP Requests, and TPM usage changes. 

Thank you!